+++ lib/utils/typedefs.dart
@@ -1,4 +1,4 @@
-import 'package:fl_language_picker/languages.dart';
+import 'package:fl_language_picker/language_locale.dart';
 import 'package:flutter/material.dart';
 
 typedef Widget ItemBuilder(Language language);